pyproject.toml
[project]
name = "pdf-chatbot-adk"
version = "1.0.0"
description = "Professional PDF Chatbot using Google ADK, OpenAI, and ChromaDB"
requires-python = ">=3.10"
dependencies = [
    "google-adk>=0.1.0",
    "openai>=1.0.0",
    "chromadb>=0.4.0",
    "streamlit>=1.30.0",
    "pypdf>=3.0.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "python-dotenv>=1.0.0",
    "tiktoken>=0.5.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.uv]
dev-dependencies = [
    "pytest>=7.0.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
]
